requirements.txt
aiogram==3.20.0
python-dotenv==1.1.0
yoomoney==0.1.2
requests